Switch to Atom as my default editor
There is this cool new editor from GitHub called Atom. It appears to
be nearly as configurable as Emacs and a whole lot prettier. So I'm
switching to Atom as my default editor.

This change also allows for Zsh init scripts to be disabled by placing
`.disabled` (or anything after the `.sh` really) at the end of the file
name.