Rewrite Atom stylesheet to be more clear
Previously, everything was written to make the LESS code "efficient" as
far as space went. But this inhibited comprehension when changes needed
to be made. It also made specificity hard to diagnose. This new approach
changes things so that each modification to Atom's style is a logical
chunk that can be turned on, turned off or tweaked indivdually.